x.crypto.chacha20: add a check counter overflow to set_counter for standard mode

+++ vlib/x/crypto/chacha20/chacha_64bitctr_test.v
@@ -109,7 +109,7 @@ fn test_chacha20_encrypt_with_64bit_counter() ! {
 }
 
 struct Test64BitCounter {
-	count      int
+	count      u64
 	key        string
 	nonce      string
 	plaintext  string
